+++ src/index.ts
@@ -15,3 +15,4 @@ export * from "./form/file-field";
 export * from "./form/date-picker";
 export * from "./form/date-field";
 export * from "./form/time-picker";
+export * from "./form/time-field";