+++ CHANGELOG.md
@@ -18,7 +18,7 @@
 -   `1.8.0`
 
     -   Added [TimePicker](src/form/time-picker/usage.md) widget
-    -   Added [TimeField](docs/time-field.md) widget
+    -   Added [TimeField](src/form/time-field/usage.md) widget
     -   Modified [SelectField](src/form/select-field/usage.md) to make option labels `inline`
     -   Added [useWindowBreakpoints](docs/use-window-breakpoints.md) hook